requirements.txt
youtube-transcript-api
deepmultilingualpunctuation
google-genai
python-dotenv
